+++ apt-pkg/contrib/mmap.h
@@ -93,8 +93,8 @@ class DynamicMMap : public MMap
    // Allocation
    unsigned long RawAllocate(unsigned long Size,unsigned long Aln = 0);
    unsigned long Allocate(unsigned long ItemSize);
-   unsigned long WriteString(const char *String,unsigned long Len = (unsigned long)-1);
-   inline unsigned long WriteString(string S) {return WriteString(S.c_str(),S.length());};
+   size_t WriteString(const char *String,size_t Len = (size_t)-1);
+   inline size_t WriteString(string S) {return WriteString(S.c_str(),S.length());};
    void UsePools(Pool &P,unsigned int Count) {Pools = &P; PoolCount = Count;};
    
    DynamicMMap(FileFd &F,unsigned long Flags,unsigned long WorkSpace = 2*1024*1024);
