+++ methods/ftp.h
@@ -13,7 +13,7 @@
 class FTPConn
 {
    char Buffer[1024*10];
-   unsigned long Len;
+   size_t Len;
    int ServerFd;
    int DataFd;
    int DataListenFd;
@@ -53,7 +53,7 @@ class FTPConn
    bool ExtGoPasv();
    
    // Query
-   bool Size(const char *Path,unsigned long &Size);
+   bool Size(const char *Path,size_t &Size);
    bool ModTime(const char *Path, time_t &Time);
    bool Get(const char *Path,FileFd &To,unsigned long Resume,
 	    Hashes &MD5,bool &Missing);
